add constraint recipe

+++ cookSort.py
@@ -25,9 +25,9 @@ except:
 db = client["childs"]
 
 def is_signed_food(data):
-    body = "{" + f'"user":"{data["user"]}","food":"{data["food"]}","like":{"true" if data['like'] else "false"}' + "}"
-    if time.time() - data["time"] < 120:
-        signature = hashlib.sha256((os.getenv('APP_SECRET') + body + f'{data["time"]}').encode()).hexdigest()
+    body = "{" + f'"user":"{data.get("user")}","food":"{data.get("food")}","like":{"true" if data.get('like') else "false"}' + "}"
+    if time.time() - data.get("time") < 120:
+        signature = hashlib.sha256((os.getenv('APP_SECRET') + body + f'{data.get("time")}').encode()).hexdigest()
         if signature == data.get("code"):
             return True
         else:
@@ -36,9 +36,35 @@ def is_signed_food(data):
     else:
         print('Later')
         return False
+def is_signed_removed(data):
+    body = "{" + f'"name":"{data.get("name")}"' + "}"
+    if time.time() - data.get("time") < 120:
+        signature = hashlib.sha256((os.getenv("APP_SECRET") + body + f'{data.get("time")}').encode()).hexdigest()
+        if signature == data.get("code"):
+            return True
+        else:
+            print("wrong code")
+            return False
+    else:
+        print("Later")
+        return False
+
 @app.route("/generate", methods=["GET"], strict_slashes=False)
 def gen():
-    return co.chat(message=f'Find a recipe (response in {flask.request.args.get("lang")})').text, 201
+    all_disliked = []
+    all_liked = []
+    collections = db.list_collection_names()
+    for collection in collections:
+        disliked = list(db[collection].find({"like": False}))
+        liked = list(db[collection].find({"like": True}))
+        for e in disliked:
+            if e["food"] not in all_disliked:
+                all_disliked.append(e["food"])
+        for e in liked:
+            if e["food"] not in all_liked:
+                all_liked.append(e["food"])
+    print(all_liked, all_disliked)
+    return co.chat(message=f"Find a recipe for {len(collections)} people (answer in {flask.request.args.get("lang")}, without : {all_disliked} and try to prioritize : {all_liked})").text, 201
 @app.route("/addfood", methods=["POST"], strict_slashes=False)
 def addfood():
     data = flask.request.get_json()
@@ -61,12 +87,16 @@ def addfood():
 @app.route('/removechild', methods=["POST"], strict_slashes=False)
 def removechild():
     data = flask.request.get_json()
-    if data.get("name") in db.list_collection_names():
-        db[data.get("name")].drop()
-        return f'{data.get("name")} was deleted', 201
+    if is_signed_removed(data):
+        if data.get("name") in db.list_collection_names():
+            db[data.get("name")].drop()
+            return f'{data.get("name")} was deleted', 201
+        else:
+            print("Can't remove child")
+            return "Can't remove child", 401
     else:
-        print("Can't remove child")
-        return "Can't remove child", 401
+        print("Can't removed this child")
+        return "Can't removed this child", 201
 
 if __name__ == "__main__":
     try:
